Add files via upload
At the moment the api data is a template that gets loaded from the Homepage.js and get passed to the equipmentTable.js.  At the moment there is no code for when it doesn't get the api data (has I have no api to test) but this will be put in closer when it is getting finished.